Automation script fixes
